backend/requirements.txt
flask
tensorflow
pillow
numpy
pandas
scikit-learn